Refactor `InstrumentationContext`
